# api/openapi.yaml
openapi: 3.0.3
info:
    title: Distributed RandomForest API
    description: API for training and inference of distributed RandomForest models.
    version: 1.0.0

paths:
    /train:
      post:
        summary: Start a new training process
        description: Start the asynchronous training of a new RandomForest model.
        requestBoody:
          required: true
          content:
            application/json:
              schema:
                type:object
              properties:
                dataset_url:
                  type: string
                  description: public URL of the training dataset.
                  example: "s3.//my-datasets/iris.csv"
                task_type:
                  type: string
                  enum: [ classification, regression ]
                  description: Type of ML problem to solve.
                # (examples of hyperparameters)
                n_estimators:
                  type: integer
                  description: Total number of trees to train in the forest.
                  default: 100
                maxdepth:
                  type: integer
                  description: Maximum nesting depth of each tree.
                  nullable: true

              required: dataset_url
        responses:
          '202':
            description: Training accepted and started in background.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    model_id:
                      type: string
                      format: uuid
                      description: unique ID assigned to the model for future requests.
          '400':
            description: Invalid request (eg. malformed URL)

    /models/{model_id}:
      get:
        summary: Check model status
        parameters:
          - name: model_id
            description: Unique identifier of the trained model.
            in: path
            required: true
            schema:
              type: string
              format: uuid
        responses:
          '200':
            description: Current state of the model.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    model_id:
                      type: string
                      format: uuid
                    status:
                      type: string
                      enum: [ training, ready, failed ]
                    created_at:
                      type: string
                      format: date-time
          '404':
            description: Model not found.

    /predict/{model_id}:
      post:
        summary: Execute inference using a trained model
        parameters:
          - name: model_id
            in: path
            description: Unique identifier of the trained model.
            required: true
            schema:
              type: string
              format: uuid
        requestBody:
          description: Data to use for the prediction. Format must be a JSON object where each key is the name of a feature and the value is the corresponding feature value.
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  features:
                    type: array
                    items:
                      type: number
                    description: Array of feature values
                  task_type:
                    type: string
                    enum: [ classification, regression ]
                    description: The type of the model (must match training task type)
                required:
                  - features
                  - task_type
        responses:
          '200':
            description: Prediction result.
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    prediction:
                      # todo enter prediction format
                      oneOf:
                        - type: string
                        - type: number
          '400':
            description: Invalid inference input.
          '404':
            description: Model not found.
          '409':
            description: Model not yet ready for inference (still training).